chore: split cells into progress, plan, terminal sub-modules
cells/mod.rs: 1565 → 905. Extracted:
- progress.rs: ProgressRole + push_progress_group/push_live_events (266 lines)
- plan.rs: parse_render_plan_block + compact_live_response (237 lines)
- terminal.rs: terminal_cell_from_entries + parse helpers (185 lines)

Add pub(super) qualifiers on extracted functions and fix cross-module imports.
