+++ src/tui/render/cells/mod.rs
@@ -78,672 +78,25 @@ fn is_progress_stack_title(line: &Line<'static>) -> bool {
     )
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(super) enum ProgressRole {
-    Thinking,
-    Exploring,
-    Planning,
-    Running,
-}
-
-impl ProgressRole {
-    fn from_entry_role(role: &str) -> Option<Self> {
-        match role {
-            "Thinking" => Some(Self::Thinking),
-            "Exploring" => Some(Self::Exploring),
-            "Planning" => Some(Self::Planning),
-            "Running" => Some(Self::Running),
-            _ => None,
-        }
-    }
-
-    fn from_live_event(event: &ActiveLiveEvent) -> Self {
-        match event {
-            ActiveLiveEvent::Thinking(_) => Self::Thinking,
-            ActiveLiveEvent::ExplorationAction(_) | ActiveLiveEvent::ExplorationNote(_) => {
-                Self::Exploring
-            }
-            ActiveLiveEvent::PlanningAction(_) | ActiveLiveEvent::PlanningNote(_) => Self::Planning,
-            ActiveLiveEvent::RunningAction(_) => Self::Running,
-        }
-    }
-}
-
-fn progress_entry_message_lines(role: ProgressRole, message: &str) -> Vec<String> {
-    match role {
-        ProgressRole::Thinking => message
-            .lines()
-            .filter(|line| !line.trim().is_empty())
-            .map(ToString::to_string)
-            .collect(),
-        ProgressRole::Exploring | ProgressRole::Planning | ProgressRole::Running => message
-            .lines()
-            .map(str::trim)
-            .filter(|line| !line.is_empty())
-            .map(|line| {
-                line.trim_start_matches("└")
-                    .trim_start_matches('•')
-                    .trim()
-                    .to_string()
-            })
-            .filter(|line| !line.is_empty())
-            .collect(),
-    }
-}
-
-fn explicit_progress_entry_groups<'a>(
-    entries: impl Iterator<Item = &'a TranscriptEntry>,
-) -> Vec<(ProgressRole, Vec<String>)> {
-    let mut groups: Vec<(ProgressRole, Vec<String>)> = Vec::new();
-    for entry in entries {
-        let Some(role) = ProgressRole::from_entry_role(entry.role.as_str()) else {
-            continue;
-        };
-        let messages = progress_entry_message_lines(role, &entry.message);
-        if messages.is_empty() {
-            continue;
-        }
-        if let Some((last_role, last_messages)) = groups.last_mut()
-            && *last_role == role
-        {
-            last_messages.extend(messages);
-            continue;
-        }
-        groups.push((role, messages));
-    }
-    groups
-}
-
-fn push_progress_group<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    role: ProgressRole,
-    messages: Vec<String>,
-    active: bool,
-) {
-    match role {
-        ProgressRole::Thinking => {
-            cells.push(Box::new(ThinkingTextCell::new(&messages.join("\n"), 4)))
-        }
-        ProgressRole::Exploring => cells.push(Box::new(ExploringCell::new(
-            compact_summary_lines(messages.as_slice(), 4, "more exploration step(s)"),
-            active,
-        ))),
-        ProgressRole::Planning => cells.push(Box::new(PlanningCell::new(
-            compact_summary_lines(messages.as_slice(), 4, "more planning step(s)"),
-            active,
-        ))),
-        ProgressRole::Running => cells.push(Box::new(RunningCell::new(
-            compact_summary_lines(messages.as_slice(), 4, "more running step(s)"),
-            active,
-        ))),
-    }
-}
-
-fn push_live_events<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    events: &[crate::tui::state::ActiveLiveEvent],
-    streaming_thinking_lines: Option<&'a [Line<'static>]>,
-    active: bool,
-) {
-    let mut thinking_messages = Vec::new();
-    let mut exploration_actions = Vec::new();
-    let mut exploration_notes = Vec::new();
-    let mut planning_actions = Vec::new();
-    let mut planning_notes = Vec::new();
-    let mut running_actions = Vec::new();
-
-    for event in events {
-        match ProgressRole::from_live_event(event) {
-            ProgressRole::Thinking => {
-                push_live_exploration_group(
-                    cells,
-                    &mut exploration_actions,
-                    &mut exploration_notes,
-                    active,
-                );
-                push_live_planning_group(cells, &mut planning_actions, &mut planning_notes, active);
-                push_live_running_group(cells, &mut running_actions, active);
-                thinking_messages.push(event.message().to_string());
-            }
-            ProgressRole::Exploring => {
-                push_live_thinking_group(cells, &mut thinking_messages, None);
-                push_live_planning_group(cells, &mut planning_actions, &mut planning_notes, active);
-                push_live_running_group(cells, &mut running_actions, active);
-                if event.is_note() {
-                    exploration_notes.push(event.message().to_string());
-                } else {
-                    exploration_actions.push(event.message().to_string());
-                }
-            }
-            ProgressRole::Planning => {
-                push_live_thinking_group(cells, &mut thinking_messages, None);
-                push_live_exploration_group(
-                    cells,
-                    &mut exploration_actions,
-                    &mut exploration_notes,
-                    active,
-                );
-                push_live_running_group(cells, &mut running_actions, active);
-                if event.is_note() {
-                    planning_notes.push(event.message().to_string());
-                } else {
-                    planning_actions.push(event.message().to_string());
-                }
-            }
-            ProgressRole::Running => {
-                push_live_thinking_group(cells, &mut thinking_messages, None);
-                push_live_exploration_group(
-                    cells,
-                    &mut exploration_actions,
-                    &mut exploration_notes,
-                    active,
-                );
-                push_live_planning_group(cells, &mut planning_actions, &mut planning_notes, active);
-                running_actions.push(event.message().to_string());
-            }
-        }
-    }
-
-    push_live_exploration_group(
-        cells,
-        &mut exploration_actions,
-        &mut exploration_notes,
-        active,
-    );
-    push_live_planning_group(cells, &mut planning_actions, &mut planning_notes, active);
-    push_live_running_group(cells, &mut running_actions, active);
-    push_live_thinking_group(cells, &mut thinking_messages, streaming_thinking_lines);
-}
-
-fn push_live_thinking_group<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    messages: &mut Vec<String>,
-    stream_lines: Option<&'a [Line<'static>]>,
-) {
-    if messages.is_empty() && stream_lines.map_or(true, |lines| lines.is_empty()) {
-        return;
-    }
-    if stream_lines.is_some() {
-        cells.push(Box::new(ThinkingGroupCell::new(
-            std::mem::take(messages).join("\n"),
-            stream_lines,
-            4,
-        )));
-        return;
-    }
-    cells.push(Box::new(ThinkingTextCell::new(&messages.join("\n"), 4)));
-    messages.clear();
-}
-
-fn push_live_exploration_group<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    actions: &mut Vec<String>,
-    notes: &mut Vec<String>,
-    active: bool,
-) {
-    if actions.is_empty() && notes.is_empty() {
-        return;
-    }
-    cells.push(Box::new(ExploringCell::new(
-        compact_progress_summary_lines(
-            actions.as_slice(),
-            notes.as_slice(),
-            4,
-            "more exploration step(s)",
-        ),
-        active,
-    )));
-    actions.clear();
-    notes.clear();
-}
-
-fn push_live_planning_group<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    actions: &mut Vec<String>,
-    notes: &mut Vec<String>,
-    active: bool,
-) {
-    if actions.is_empty() && notes.is_empty() {
-        return;
-    }
-    cells.push(Box::new(PlanningCell::new(
-        compact_progress_summary_lines(
-            actions.as_slice(),
-            notes.as_slice(),
-            4,
-            "more planning step(s)",
-        ),
-        active,
-    )));
-    actions.clear();
-    notes.clear();
-}
-
-fn push_live_running_group<'a>(
-    cells: &mut Vec<Box<dyn HistoryCell + 'a>>,
-    actions: &mut Vec<String>,
-    active: bool,
-) {
-    if actions.is_empty() {
-        return;
-    }
-    cells.push(Box::new(RunningCell::new(
-        compact_recent_first_summary_lines(actions.as_slice(), 4, "more running step(s)"),
-        active,
-    )));
-    actions.clear();
-}
-
-fn split_progress_sentences(message: &str) -> Vec<String> {
-    let mut sentences = Vec::new();
-    let mut current = String::new();
-    let mut chars = message.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        current.push(ch);
-
-        let next = chars.peek().copied();
-        let previous = current.chars().rev().nth(1);
-        let is_decimal_separator = ch == '.'
-            && previous.is_some_and(|prev| prev.is_ascii_digit())
-            && next.is_some_and(|peek| peek.is_ascii_digit());
-        let continues_punctuation = next.is_some_and(|peek| matches!(peek, '.' | '!' | '?'));
-
-        if matches!(ch, '.' | '!' | '?') && !is_decimal_separator && !continues_punctuation {
-            let trimmed = current.trim();
-            if !trimmed.is_empty() {
-                sentences.push(trimmed.to_string());
-            }
-            current.clear();
-        }
-    }
-
-    let tail = current.trim();
-    if !tail.is_empty() {
-        sentences.push(tail.to_string());
-    }
-
-    sentences
-}
-
-fn is_structured_response_marker(line: &str) -> bool {
-    let trimmed = line.trim();
-    trimmed.starts_with("<proposed_plan>")
-        || trimmed.starts_with("</proposed_plan>")
-        || trimmed.starts_with("<plan>")
-        || trimmed.starts_with("</plan>")
-        || trimmed.starts_with("<request_user_input>")
-        || trimmed.starts_with("</request_user_input>")
-        || trimmed.starts_with("<continue_inspection")
-}
-
-fn is_structured_progress_list_line(line: &str) -> bool {
-    let trimmed = line.trim_start();
-    trimmed.starts_with("- [")
-        || trimmed.starts_with("* [")
-        || trimmed.starts_with("• [")
-        || trimmed.starts_with("- ")
-        || trimmed.starts_with("* ")
-        || trimmed.starts_with("• ")
-}
-
-fn compact_live_response_source(message: &str) -> Option<String> {
-    let mut retained = Vec::new();
-    let mut saw_prose = false;
-    let mut in_structured_block = false;
-
-    for line in message.lines() {
-        let trimmed = line.trim();
-        if is_structured_response_marker(trimmed) {
-            if trimmed.starts_with("</") {
-                in_structured_block = false;
-            } else if trimmed.starts_with('<') && trimmed.ends_with('>') && !trimmed.ends_with("/>")
-            {
-                in_structured_block = true;
-            }
-            continue;
-        }
-
-        if in_structured_block {
-            continue;
-        }
-
-        if trimmed.is_empty() {
-            if saw_prose {
-                retained.push(String::new());
-            }
-            continue;
-        }
-
-        if is_structured_progress_list_line(trimmed) && saw_prose {
-            continue;
-        }
-
-        retained.push(trimmed.to_string());
-        saw_prose = true;
-    }
-
-    let compact = retained
-        .into_iter()
-        .skip_while(|line| line.is_empty())
-        .collect::<Vec<_>>()
-        .join("\n")
-        .trim()
-        .to_string();
-
-    if compact.is_empty() {
-        None
-    } else {
-        Some(compact)
-    }
-}
-
-fn compact_live_response_message(message: &str) -> Option<String> {
-    let source = compact_live_response_source(message)?;
-    let sentences = split_progress_sentences(&source);
-    if sentences.len() <= 3 {
-        return Some(sentences.join("\n"));
-    }
-
-    let next_markers = [
-        "next ",
-        "i will ",
-        "i'll ",
-        "then i will ",
-        "then i'll ",
-        "i am going to ",
-    ];
-
-    let mut selected_indices = vec![0];
-    let mut next_step_idx = None;
-
-    for (idx, sentence) in sentences.iter().enumerate().skip(1) {
-        let lowered = sentence.to_ascii_lowercase();
-        if next_step_idx.is_none()
-            && next_markers
-                .iter()
-                .any(|marker| lowered.starts_with(marker))
-        {
-            next_step_idx = Some(idx);
-            break;
-        }
-    }
-
-    if let Some(idx) = next_step_idx {
-        selected_indices.push(idx);
-    }
-
-    let mut idx = 1;
-    while selected_indices.len() < 3 && idx < sentences.len() {
-        if !selected_indices.contains(&idx) {
-            selected_indices.push(idx);
-        }
-        idx += 1;
-    }
-
-    selected_indices.sort_unstable();
-    Some(
-        selected_indices
-            .into_iter()
-            .map(|idx| sentences[idx].clone())
-            .collect::<Vec<_>>()
-            .join("\n"),
-    )
-}
-
-fn parse_render_plan_block(message: &str) -> Option<(Vec<(String, String)>, Option<String>)> {
-    let (start_tag, end_tag, start, end) = find_render_plan_block_bounds(message)
-        .or_else(|| find_render_legacy_plan_block_bounds(message))?;
-    if end <= start {
-        return None;
-    }
-
-    let block = &message[start + start_tag.len()..end];
-    let mut steps = Vec::new();
-    for line in block.lines().map(str::trim).filter(|line| !line.is_empty()) {
-        if let Some(step) = parse_render_plan_step_line(line) {
-            steps.push(step);
-        }
-    }
-
-    if steps.is_empty() {
-        if start_tag != "<proposed_plan>" {
-            return None;
-        }
-        let fallback = block
-            .lines()
-            .map(str::trim)
-            .find(|line| !line.is_empty() && !line.starts_with('#'))
-            .unwrap_or("Implement proposed plan")
-            .trim_matches(['*', '#', ' '])
-            .to_string();
-        steps.push(("pending".to_string(), fallback));
-    }
-
-    let explanation = message[end + end_tag.len()..].trim();
-    Some((
-        steps,
-        (!explanation.is_empty()).then(|| explanation.to_string()),
-    ))
-}
-
-fn find_render_plan_block_bounds(
-    message: &str,
-) -> Option<(&'static str, &'static str, usize, usize)> {
-    let start_tag = "<proposed_plan>";
-    let end_tag = "</proposed_plan>";
-    let start = message.find(start_tag)?;
-    let end = message.find(end_tag)?;
-    Some((start_tag, end_tag, start, end))
-}
-
-fn find_render_legacy_plan_block_bounds(
-    message: &str,
-) -> Option<(&'static str, &'static str, usize, usize)> {
-    let start_tag = "<plan>";
-    let end_tag = "</plan>";
-    let start = message.find(start_tag)?;
-    let end = message.find(end_tag)?;
-    Some((start_tag, end_tag, start, end))
-}
-
-fn parse_render_plan_step_line(line: &str) -> Option<(String, String)> {
-    if let Some(step) = line
-        .strip_prefix("- [")
-        .or_else(|| line.strip_prefix("* ["))
-        .or_else(|| line.strip_prefix("• ["))
-    {
-        let Some((status, rest)) = step.split_once(']') else {
-            return None;
-        };
-        let step = rest.trim();
-        return (!step.is_empty()).then(|| (status.trim().to_string(), step.to_string()));
-    }
-
-    let step = line
-        .strip_prefix("- ")
-        .or_else(|| line.strip_prefix("* "))
-        .or_else(|| line.strip_prefix("• "))
-        .or_else(|| {
-            let (number, rest) = line.split_once(". ")?;
-            number.chars().all(|ch| ch.is_ascii_digit()).then_some(rest)
-        })?
-        .trim();
-    (!step.is_empty()).then(|| ("pending".to_string(), step.to_string()))
-}
-
-fn terminal_cell_from_entries<'a>(
-    entries: impl DoubleEndedIterator<Item = &'a TranscriptEntry>,
-) -> Option<TerminalCell> {
-    let data = entries
-        .filter_map(terminal_cell_data_from_entry)
-        .next_back()?;
-    Some(TerminalCell::new(
-        data.command,
-        data.output,
-        data.active,
-        data.success,
-    ))
-}
-
-fn terminal_cell_data_from_entry(entry: &TranscriptEntry) -> Option<TerminalCellData> {
-    if let Some(TranscriptEntryPayload::Terminal(event)) = entry.payload.as_ref() {
-        return terminal_cell_data_from_event(&event);
-    }
-
-    if matches!(entry.role.as_str(), "Tool Result" | "Tool Error") {
-        return parse_terminal_tool_result(&entry.message);
-    }
-
-    None
-}
-
-fn terminal_cell_data_from_event(event: &TerminalEvent) -> Option<TerminalCellData> {
-    match event {
-        TerminalEvent::Begin(command) => Some(terminal_cell_data_from_command(command, true)),
-        TerminalEvent::End(command) => Some(terminal_cell_data_from_command(command, false)),
-        TerminalEvent::List(collection) => {
-            Some(terminal_cell_data_from_collection(collection, "list"))
-        }
-        TerminalEvent::Stop(collection) => {
-            Some(terminal_cell_data_from_collection(collection, "stop"))
-        }
-        TerminalEvent::OutputDelta(_) => None,
-    }
-}
-
-fn terminal_cell_data_from_command(
-    command: &TerminalCommandEvent,
-    force_active: bool,
-) -> TerminalCellData {
-    let target = match command.target {
-        TerminalTarget::Pty => "pty",
-        TerminalTarget::BackgroundTask => "background",
-    };
-    let command_label = command
-        .command
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .or(command.id.as_deref())
-        .unwrap_or("command");
-    let mut output = command.output.clone();
-    if let Some(output_path) = command.output_path.as_deref() {
-        if !output_path.trim().is_empty() {
-            output.push(output_path.to_string());
-        }
-    }
-    TerminalCellData {
-        command: format!("{target} {command_label}"),
-        output,
-        active: force_active || command.status == "running",
-        success: if force_active {
-            None
-        } else {
-            terminal_status_success(&command.status)
-        },
-    }
-}
-
-fn terminal_cell_data_from_collection(
-    collection: &TerminalCollectionEvent,
-    action: &str,
-) -> TerminalCellData {
-    let target = match collection.target {
-        TerminalTarget::Pty => "pty",
-        TerminalTarget::BackgroundTask => "background",
-    };
-    let output = collection
-        .items
-        .iter()
-        .take(6)
-        .map(|item| {
-            let id = item.id.as_deref().unwrap_or("<unknown>");
-            let command = item
-                .command
-                .as_deref()
-                .filter(|value| !value.trim().is_empty())
-                .unwrap_or("command unavailable");
-            format!("{id} {}: {command}", item.status)
-        })
-        .collect::<Vec<_>>();
-    TerminalCellData {
-        command: format!("{target} {action}"),
-        output,
-        active: false,
-        success: Some(!collection.items.iter().any(|item| item.is_error)),
-    }
-}
-
-fn parse_terminal_tool_result(message: &str) -> Option<TerminalCellData> {
-    let mut lines = message.lines();
-    let first = lines.next()?.trim();
-
-    let mut output = Vec::new();
-    let mut in_output = false;
-    for line in lines {
-        let trimmed = line.trim();
-        if trimmed == "output:" {
-            in_output = true;
-            continue;
-        }
-        if let Some(inline_output) = trimmed.strip_prefix("output:") {
-            let inline_output = inline_output.trim();
-            if !inline_output.is_empty() {
-                output.push(inline_output.to_string());
-            }
-            in_output = true;
-            continue;
-        }
-        if in_output {
-            output.push(line.trim_end().to_string());
-        }
-    }
-
-    if let Some(rest) = first.strip_prefix("pty ") {
-        let (head, command) = parse_terminal_result_head(rest);
-        let status = head.split_whitespace().nth(1).unwrap_or("unknown");
-        return Some(TerminalCellData {
-            command: format!("pty {command}"),
-            output,
-            active: status == "running",
-            success: terminal_status_success(status),
-        });
-    }
-
-    if let Some(rest) = first.strip_prefix("background task ") {
-        let (head, command) = parse_terminal_result_head(rest);
-        let status = head.split_whitespace().nth(1).unwrap_or("unknown");
-        return Some(TerminalCellData {
-            command: format!("background {command}"),
-            output,
-            active: status == "running",
-            success: terminal_status_success(status),
-        });
-    }
-
-    None
-}
-
-fn parse_terminal_result_head(rest: &str) -> (&str, String) {
-    if let Some((head, command)) = rest.split_once(": ") {
-        return (head, command.to_string());
-    }
-    let fallback = rest
-        .split_whitespace()
-        .next()
-        .filter(|value| !value.is_empty())
-        .unwrap_or(rest);
-    (rest, fallback.to_string())
-}
-
-fn terminal_status_success(status: &str) -> Option<bool> {
-    match status {
-        "running" => None,
-        "completed" | "stopped" => Some(true),
-        "failed" | "killed" => Some(false),
-        _ => None,
-    }
-}
+pub(super) mod progress;
+use self::progress::{
+    ProgressRole, explicit_progress_entry_groups, progress_entry_message_lines, push_live_events,
+    push_live_exploration_group, push_live_planning_group, push_live_running_group,
+    push_live_thinking_group, push_progress_group,
+};
+pub(super) mod plan;
+use self::plan::{
+    compact_live_response_message, compact_live_response_source,
+    find_render_legacy_plan_block_bounds, find_render_plan_block_bounds,
+    is_structured_progress_list_line, is_structured_response_marker, parse_render_plan_block,
+    parse_render_plan_step_line, split_progress_sentences,
+};
+pub(super) mod terminal;
+use self::terminal::{
+    parse_terminal_result_head, parse_terminal_tool_result, terminal_cell_data_from_collection,
+    terminal_cell_data_from_command, terminal_cell_data_from_entry, terminal_cell_data_from_event,
+    terminal_cell_from_entries, terminal_status_success,
+};
 
 fn ordered_exploration_agent_segments<'a>(
     current_turn: &[&'a TranscriptEntry],
